rename easybuild/tools/bwraptools.py to easybuild/tools/bwrap.py + minor code style fixes
