+++ easybuild/tools/bwrap.py
@@ -40,12 +40,12 @@ from easybuild.tools.filetools import mkdir, write_file
 from easybuild.tools.utilities import trace_msg
 
 BWRAP_INFO = {
-    'modules_to_install': set(),
-    'installpath_software': '',
-    'installpath_modules': '',
-    'bwrap_installpath': '',
     'bwrap_cmd': [],
     'bwrap_eb_options': [],
+    'bwrap_installpath': '',
+    'installpath_modules': '',
+    'installpath_software': '',
+    'modules_to_install': set(),
 
 }
 BWRAP_INFO_JSON = 'bwrap_info.json'
@@ -58,30 +58,32 @@ def prepare_bwrap(bwrap_installpath):
     Prepare for running EasyBuild with bwrap:
     - update BWRAP_INFO
     - write json metadata file with BWRAP_INFO
-    - set environment variable EB_BWRAP_CMD
+    - set environment variable $EB_BWRAP_CMD
 
-    :param bwrap_installpath: bwrap installpath
+    :param bwrap_installpath: bwrap install path
     """
 
     BWRAP_INFO['bwrap_installpath'] = bwrap_installpath
     BWRAP_INFO['installpath_software'] = install_path(typ='software')
     BWRAP_INFO['installpath_modules'] = install_path(typ='modules')
+
     installpath_software = BWRAP_INFO['installpath_software']
-    bwrap_installpath = BWRAP_INFO['bwrap_installpath']
-    bwrap_mpath = os.path.join(bwrap_installpath, 'modules')
+    bwrap_modules_installpath = os.path.join(bwrap_installpath, 'modules')
+
     bwrap_cmd = ['bwrap', '--dev-bind', '/', '/']
 
+    # bind mount all software directories
     for mod in BWRAP_INFO['modules_to_install']:
-        spath = os.path.join(os.path.realpath(installpath_software), mod)
-        bwrap_spath = os.path.join(bwrap_installpath, 'software', mod)
-        mkdir(spath, parents=True)
-        mkdir(bwrap_spath, parents=True)
-        bwrap_cmd.extend(['--bind', bwrap_spath, spath])
+        installdir = os.path.join(os.path.realpath(installpath_software), mod)
+        bwrap_installdir = os.path.join(bwrap_installpath, 'software', mod)
+        mkdir(installdir, parents=True)
+        mkdir(bwrap_installdir, parents=True)
+        bwrap_cmd.extend(['--bind', bwrap_installdir, installdir])
 
     BWRAP_INFO['bwrap_cmd'] = bwrap_cmd
 
     # disable `--bwrap` to prepare for a real installation (in bwrap namespace)
-    BWRAP_INFO['bwrap_eb_options'] = ['--disable-bwrap', f'--installpath-modules={bwrap_mpath}']
+    BWRAP_INFO['bwrap_eb_options'] = ['--disable-bwrap', f'--installpath-modules={bwrap_modules_installpath}']
 
     _log.info(f'Info needed for bwrap: {BWRAP_INFO}')
 
@@ -90,10 +92,10 @@ def prepare_bwrap(bwrap_installpath):
     write_file(bwrap_infopath, json.dumps(BWRAP_INFO, default=list, indent=2, sort_keys=True), backup=True)
 
     print_msg('Building/installing in bwrap namespace')
+    trace_msg(f'bwrap command (to prefix eb command): {" ".join(BWRAP_INFO["bwrap_cmd"])}')
     trace_msg(f'bwrap info file: {bwrap_infopath}')
     trace_msg(f'bwrap EasyBuild options: {BWRAP_INFO["bwrap_eb_options"]}')
-    trace_msg(f'bwrap prefix: {" ".join(BWRAP_INFO["bwrap_cmd"])}')
 
-    # set environment variable EB_BWRAP_CMD to make it available for the interactive debug shell
+    # set environment variable $EB_BWRAP_CMD to make it available for the interactive debug shell
     # when rerunning with bwrap
     os.environ['EB_BWRAP_CMD'] = ' '.join(BWRAP_INFO['bwrap_cmd'])
